org-id-scanner.cpp: Reset errno after a failed stat
This is because if I have a some_file opened in emacs, then there is a
./.#some_file symbolic link that points nowhere. Then when the stat on
it fails, it sets ::errno to non-zero value. After the loop, when
errno is non-zero, it fails out. So, in the case of skipping files
that could not be stat-ed, just reset the errno.